Move Conversation to Api namespace
